requirements.txt
pandas>=2.0
numpy>=1.24
pyodbc>=5.1
python-dotenv>=1.0
